fix: prevent 32-bit µs overflow in TM time-sync timestamp

+++ include/scip2/response/time_sync.h
@@ -31,7 +31,7 @@ namespace scip2
 class Timestamp
 {
 public:
-  uint32_t timestamp_;
+  uint64_t timestamp_;
   bool is_wall_;
 
   Timestamp()
